removed i18nConfig and replaced with settings in application.properties and beans in DrivedokApplication. Added two img and a language chooser in layout. This does need some more work though
