+++ src/main/resources/application.properties
@@ -3,3 +3,6 @@ server.error.whitelabel.enabled=false
 server.error.include-exception=true
 server.error.include-stacktrace=always
 spring.mvc.hiddenmethod.filter.enabled=true
+spring.messages.basename=il8n/messages
+spring.messages.encoding=UTF-8
+spring.messages.use-code-as-default-message=true
