+++ src/main/resources/il8n/messages.properties
@@ -2,6 +2,8 @@
 btn.save=Save
 btn.reset=Reset
 btn.delete=Delete
+main.language.nl=Dutch
+main.language.en=English
 
 # reservation-listpage
 reservation.listpage.title=Reservations
